+++ tests/commands/command_test.py
@@ -80,6 +80,7 @@ class CommandTest(CmdLineTest):
         config.commands.open.command = "cat"
         config.database.file = str(self.COBIB_TEST_DIR / "database.yaml")
         config.database.git = request.param.get("git", False)
+        config.utils.file_downloader.default_location = "/tmp"
 
         # load database
         if request.param.get("database", True):